CIV-9406 Trial Ready document owner assigned using information retrived from case data.
